i40iw: Add Quality of Service support
Add support to assign QoS traffic class (TC) to i40iw iwarp traffic flows.
During device initialization, a qos table with the mapping of user
priority (0-7) to a TC's queue set handle is provided by the i40e LAN
driver. A QP has its user priority assigned depending on the ToS
requested in the connection setup. The queue set handle associated with
the QP's user priority is looked up from the qos table to program the
QP context and assign the TC for the QP.

Signed-off-by: Faisal Latif <faisal.latif@intel.com>
Signed-off-by: Shiraz Saleem <shiraz.saleem@intel.com>
Signed-off-by: Henry Orosco <henry.orosco@intel.com>